refactor: extract theme_picker / pin_duration / poll_vote overlays (#357)
Seventh slice. Three medium overlays bundled in one PR:

- draw_theme_picker         (~88 lines) - theme list with colour swatches
- draw_pin_duration_picker  (~32 lines) - PIN_DURATIONS choice list
- draw_poll_vote_overlay    (~57 lines) - multi-select vote checkboxes

PIN_DURATIONS is now only used by the pin_duration overlay, so it was
dropped from ui/mod.rs's import block.

ui/mod.rs: 4610 -> 4434 lines.
